Think more about query language
After reading up more on Datalog, a tuple-centric query language makes a
lot of sense. I should not bother with operators, just expand out
everything into 3-tuples of (entity, attribute, value). A statement is
true if such a tuple exists in the database. Equalities are still
possible, you just have to make an intermediate variable and then add
the two statements.

For example, to join "build" with "push" on "commit", rather than an
expression like

    build.commit b == push.commit p

we write two statements

    b build.commit commit
    p push.commit commit

With this approach, the implementation also became a lot clearer to me:
nested loops, one for each free entity variable. If the value of a free
variable is implied by a bound variable (bound in a surrounding loop),
then we might replace the loop with a non-loop. Cardinality constraints
can also help. For instance, with

    p project.company c
    c company.name name

we would in principle have to iterate over the cartesian product of
projects and companies. But if the "project.company" cardinality is
single (the default), then we know that only one (p, project.company, c)
tuple can exist (at a time, previous tuples must have been retracted),
and c is no longer a free variable; it is implied by the project. If we
would iterate over companies in the outer loop, then there is no such
relationship that limits the project, although the avet index could be
used to find all projects quickly.