Add method to assert query plan valitity
If this thing is going to handle serious data, I had better put in lots
of debug checks everywhere. Asserting that a query plan is well-typed,
and that the invariant about the variables is respected, seems like a
useful thing to check, because I am surely going to get this wrong.

It would be nice to have GADTs to tag variables with a type, although
storing them in a collection then becomes difficult because the
collection is then heterogeneous.