Begin defining query plan evaluator
It will be a set of iterators, one for every loop. Increment the last
one, and once it is exhausted, increment the one before and then reset
the last one, etc.

A nice side effect is that by construction, all results can be streamed.
Although when user-requested ordering is introduced that would change.