Fix bug in query plan evaluation
It now does not loop infinitely any more. The query now does not return
any results, but that is correct; I request all attributes, but the
attributes "many" and "unique" are not set on all attributes, and in
particular, they are never set both at the same time (currently). So
that means that for no assignment of variables, all of the tuples in the
query are present in the database. The query returns no results, and
that is correct.

This does force me to reconsider nullable data. I think this is a good
thing. Previously I was not sure whether to add booleans for things like
"unique" and "many", or to add a unit type, and use the presence or
absence of an attribute as an indication. Now, it is clear to me that I
need to use booleans and always set the attribute.

How would one deal with the nullable case though, I can imagine that it
is useful. Introducing an explicit "or attribute is not set" in the
query is an option. This way nullability becomes explicit. By default
queries never return null values; every result where one of the values
would be null, would not be a valid result. Also good to remember:
putting in null is an explicit choice. The default is not nullability.
And that is good. The real question is, how to deal with optional
attributes. Letting the user decide at select time is not such a bad
idea. If you select an attribute, then the result only contains entities
that have that attribute. If it's optional, then don't select it. Only
in the case where you want both in the same query, you need a fallback.
And I think that for now, I can manage without that.