Add a very naive query planner
It plans every variable the first time it sees it, in the order
variables occur in the query. It does not rearrange to change the loop
order. Also, it does not cover all possible cases; it can only compute
cartesian products and variables defined by one-sided implications
(where either entity and attribute are known, or attribute and value),
but it does not filter if there are multiple restrictions on a variable.

For now this will do.