Think about the disk format
I want to straighten out how to represent the B+ trees on disk before
writing any code. I have a rough idea of how to do it for one immutable
tree, and also how to merge trees, but I am not yet sure how to
partition ranges over multiple files, to ensure that we don't have to
scan *all* of the files.

I have the feeling that if I think about it, it will be straightforward.
Writing things down helps to sort this out. And before I can dive into
the low-level details, it is good to write down the high-level idea.