Think about log-structured merge trees
Now I'm thinking about this, they are trees of trees ... is it not just
a trick to keep the tree balanced then? Feels very similar to hitchhiker
trees. Thinking about something simple without complex balancing logic,
I arrived at a chain of trees which can do log(n) insert, but query is
more expensive. Or perhaps I made a mistake in the computation and it is
worse. Will need to think more about this.