Add alignment annotations
We don't really *need* datoms to be 32-byte algined, but I'd like them
to. In practice the ones in a tree node are, and it might be helpful to
inform the compiler about this. It will also cause datoms that live
elsewhere to be aligned to 32 bytes, but that is not a bad idea anyway,
because it means that a datom will not straddle a cache line.