Rethink node memory and disk layout
I think it might make sense to fix the number of slots, and not split
the datom array into pending and midpoint datoms. The trade off is as
follows: we pay a bit in space: we can fit at most 101 datoms in a page,
as opposed to 127 for leaves of pending datoms. The child array will
always be there, but some slots will be unused to indicate pending
datoms.

What we get in return is that we can merge the pending and midpoint
datoms into one array, and sort it. This simplifies traversal and should
also help performance a bit, because it is no longer necessary to do a
merge sort upon traversal. If the B-tree hypothesis (compares are cheap,
disk access is slow) is true, then this is not a good trade off. But I
would like to make it anyway, at least for now, for the sake of
simplicity.