Store upper bound in nodes itself
Rather than giving nodes with midpoints m0..mk a layout like so:

  (-infty, m0] (m0, m1] (m1, m2] ... (mk-1, mk] (mk, +infty)

We give it a layout like so:

  (-infty, m0] (m0, m1] (m1, m2] ... (mk-1, mk]

With the upper bound stored in the node itself. As is clear from the
illustration, this makes the nodes more regular. It does reduce the
branching factor by one, but a branching factor of 102 vs 103 is not
that big of a deal, and it simplifies the implementation a lot. I am
willing to make the sacrifice and waste those 8 bytes for the sake of
simplicity.