Add link to unofficial Datomic internals guide
No https unfortunately, looks like it is hosted on Github pages; the
server presents itself with a *.github.com certificate.

There are some interesting conclusions to draw from this article. The
first one is that the VAET index appears to be important, and the use of
AVET is limited. But from my experiments, I could always use AVET and I
never needed VAET. This is because the attribute is fixed in the query
anyway, so it doesn't really matter which one to use: once you have an
attribute and a value, you can use either index to find the entity. I
was thinking of getting rid of VAET, actually, because I never used it,
and it is redundant if you always know both the attribute and value.
There were some caveats to what is stored in which index in Datomic
though, with VAET storing only reference values. So it will be smaller,
and faster to query. It does not change the complexity, but in practice
it might matter a lot. I don't know, but for now I see no reason for
keeping VAET.

The second notable observation is that the B-trees use a 1000+ branching
factor. I don't know if the author verified this, or if it is just a
quote of how people say B-trees are supposed to be used. Because so far
I have been able to fit only 102 datoms on a 4k page, and with some
squeezing (using 32-bit ids mainly) you could get a factor 2 or maybe 3
more, but not easily a factor 10. The article also mentions having trees
at most 3 levels deep. With a branching factor of 100, you can get 1M
datoms at 3 levels, which is probably not enough for large databases,
while 1G datoms apparently is. So we'll have to see if my current design
with 4k pages is a good one; perhaps nodes need to be much larger,
perhaps not.

Another interesting note is compression. I considered this but opted to
not do it mainly for simplicity and also for speed (a premature
optimization that is not based on measurements). I would expect index
nodes to compress extremely well: I use 64-bit entity ids and value
pointers, but for most purposes 32 bits is plenty, so there will be lots
of zeros. Furthermore, datoms are sorted, so the values are similar too,
so with a delta encoding there would be a lot to gain in terms of
storage. The big downside of compression is that you can no longer mmap
pages. Mmap has many advantages: it is fast; you can map more than
physical RAM and the OS will handle what needs to be resident; OS can
evict pages that are backed by disk in case of memory pressure; the OS
can detect patterns and prefetch; data can be shared among multiple
processes on the same machine, ... If you compress nodes then you lose
all of this, and you have to implement everything that the OS could do
for you: loading pages from disk, deciding which ones to keep in case of
memory pressure, or deciding which ones to keep at all (the app does not
have a global view like the OS does), etc. In the end I think it would
be much much simpler to use a file system with transparent compression.
And if you then want to share pages over the network, you can still
compress on the fly if so desired. (Although sendfile() is very tempting
for that too.) In short, while compression could be a big win for
storage space, in Noblit I will not compress, so I can using mmap to
sidestep many complications.