Implement htree iteration
This should now visit child nodes properly, and also merge in the
pending datoms along the way. It took a while to arrive at this design,
but now I am quite happy with it: this design is quite simple, with few
edge cases. The code is long, but some branches are entirely comments
and assertions. The actual traversal algorithm is quite simple.

A brief overview: at every level in the tree, we keep track of which
datom there to yield, in a stack, the root node at the bottom. On
next(), we look at the datom at all of these levels, and yield the
minimal one. Then we advance the index at that level. If we did not
yield from the node at the top of the stack, then we yielded a pending
datom from a parent. (It could not have been a midpoint, because the
midpoint is greater than the datoms in the child nodes.) If we yielded
from the node at the top of the stack, then we may have yielded a
midpoint. If we did, then we now need to look for the next midpoint and
dive all the way until a leave node, because the least child is stored
in a leaf. It may also be that the node is exhaustet, in which case we
pop it from the stack, such that we can yield the midpoint from the
parent next. That's it, really.