Rename Node::depth to Node::level
The term "depth" may be a bit misleading, because I used the value 0 at
the leaves, and greater values for parents, with the root having the
greatest depth. This is contrary to common terminology, where the leaves
are "deeper" in the tree. There is a good reason for it though: it means
that we can share the same leaf nodes between trees of different height.

The term "level" has less of an intrinsic preference for one direction
(unlike e.g. "height"), so it should encourage people more to not make
assumptions and check the documentation, even though people reading
documentation or verifying assumiptions is probably wishful thinking on
my part.