Write a bit of the splitting logic
This is becoming too complex. I think I will change to a B+ tree rather
than a B-tree, duplicating all node maxima as their parent midpoints. It
takes a bit more space, but it simplifies node splitting, and it should
not affect iteration too much.

Furthermore, using a B+ tree that does not need to extract the maxima,
means that we can split a node without having to rewrite nodes all the
way to level 0. That rewrite might not have been so bad because we could
do a flush at the same time, but still. But in any case, optimal
efficiency -- either in space or in insert or traversal time -- is not
what I am aiming for. I should make it work first, and only then make it
fast.