Write more of the splitting logic
I just realized -- or maybe I am completely wrong at this point and
tomorrow I wonder what exactly I was thinking -- that splitting does not
need to return two nodes and two upper bounds, only two nodes and one
midpoint. If we are splitting a node which was not the root, then we
will already have a midpoint datom pointing to it, and that will be a
valid upper bound to reuse for the right node. The only extraction that
happens is to extract the midpoint datom that will now be an upper bound
for the left part of the split. And this does not require deep rewrites,
we can just return it directly. This is a lot simpler than the max
extraction I thought I had to do previously.

Finally, there is the case of splitting the root node. There it *is*
necessary to extract the max from the right-hand side. At this point I
do not see any way of getting around that. But at least there is also a
simpler case of the split. Which then means that split is even more
complex than I thought, because the complex case is still there, only
now we additionally have a less complex case. But let's take it one step
at a time, and fill out the simpler cases first.