Make Store::write_page take a slice, not callback
This is required to make the memory store pass the borrow checker. It is
possible to have nodes, which reference memory in the current store.
When we then split that node, we still get references into that node.
And when we then try to write those, we can't mutate the store, because
there are still immutable references to it. For an on-disk store that
makes the promise that it will never mutate already-written bytes, this
would be a safe thing to do. But for the vector-backed memory store,
writing can cause the vector to reallocate, and that would invalidate
existing nodes. Rust's borrow checker truly prevented a memory bug!

Now, instead of writing to the underlying writer directly, we will first
build up the page in memory, and only then write it. This may sound bad,
but in fact it is just moving the buffering. Previously we would have to
wrap the writer in a buffered writer. Now we should *not* wrap it,
because we already do the buffering manually.