Add note about extracting midpoints
The commit is small, but I spent quite some time drawing trees on paper.
Extracting the max is straightforward on paper, but it is a bit tedious
to implement, especially with the current setup of a borrowed node,
where the borrow must be given up before we can write.

It is possible, but I don't like to do it. Perhaps the better solution
is to have open-ended ranges after all. I'll have to revisit whether
having an upper bound really simplifies iteration so much that it
justifies the extract for node split.