Make tree splitting take a node, not a page id
I will be splitting VecNodes, not nodes that are already in the tree,
for the need to split only arises after inserting extra datoms, and at
that point I already have a VecNode. This simplifies the code a bit, as
there is no need for the scoping to do the liftetime dance any longer.