Implement flushing in the tree
This turned out to be quite easy. The code is fairly simple,
fortunately. There are decisions to be made about when to flush and when
to split, and I suspect that different choices lead to vastly different
performance characteristics. But for now, the choice I made is to always
flush until the node fits in a page, and only when there is nothing to
be flushed, split the node.

All the pieces of the tree are coming together now, handling splits is
the last thing left. It will be a bit more difficult than I thought.
Previously I assumed that after splitting a node, it would always fit in
a page. But now that I went with the bulk insert, where you can insert
more datoms than a page's worth (which is a very nice API, especially
because it an also be used to implement flush), it may be necessary to
do multiple splits.