Make insert_into() return result as VecNode
Inserting may require a node split. And a split produces a (midpoint,
child page id) pair that needs to be inserted into the parent of the
node to split. So inserting into a given page returns, in addition to
the page id of the new page with datoms inserted, a list of (midpoint,
child page id) pairs. (It is a list because more than one split may be
required.) But a list of such pairs along with one additional page id
is exactly a node! So we can return the result as a vec node.

Then in flush, dealing with these additional midpoints is also very
elegant. They are already sorted, so we just need to insert them before
the midpoint that we are updating. So I thought we drain the old range,
and then insert the new range. But while looking through Vec's
documentation for a method to insert a range, I saw Vec::splice(), which
does both the remove and insert at once, exactly what I need. And by
splicing the right ranges, there is no case distinction at all any more
between getting only an updated page id, or also additional midpoints.

Also, in insert_into(), where I now always build a vec node with one
child, I can incrementally build this node to return, and fill it when
splitting is required. This part is not implemented yet.

Suddenly all pieces fit together so elegantly!