Assert that insert_into result has right level
Actually, the level of an insert result is not relevant, it is not used
for anything, if the implementation is correct. But it could prove to be
a valuable tool for catching bugs, and the result node has a level that
needs to be filled anyway, so let's fill it with the right level and use
it to build an extra sanity check. In release mode the insert might be
inlined, and then the level field can be eliminated entirely, because it
is never read in release mode.