Prepare tree insert_into() for making splits
If after flushing the node is still too large, we need to split it.
And at this point it is clear that with the current API, a split in two
pieces is not sufficient, it needs to be a split into an arbitrary
number of pieces. That requires a slight change to the split()
signature, but in exchange for that, the return value of split() is also
the right thing to return from insert_into(). So at this point, the only
thing left to do is to re-implement split().

The previous logic for determining the median midpoint and the like is
now redundant. When we get to splitting, either the node is a leaf and
so we don't need to skip over pending datoms, because a leaf has only
pending datoms. Or the node is an internal node, but then we finished
flushing, and so there are no pending datoms left, and every datom is a
midpoint. So to decide where to split, we can just divide the datoms up
over enough pages to fit all of them.

There are some policy decisions to be made about how many pages to split
a node in. For now I went with the minimum (almost minimum, see the
caveat in the comment), but that can be a bad choice if it results in
full or almost full pages after the split, because these will fill up
immediately the next time we try to insert something into them. I think
it is better to create nodes with some space for pending datoms, but
then the question is *how much* pending space. So for now I am just
going with something that is easy to compute and which will work.