Add fuzzer to fuzz insertion
This is a very basic fuzzer to start with.

 * No assertions apart from those already there.
 * Inserting only single-element slices (to avoid sorting issues).
 * Inserting only 8-bit integers.

Nonetheless, the coverage already grows very large. Even with such a
limited scenario -- inserting single elements -- there is a lot of
variety. Different permutations of insertion order might lead to
different trees (because of the "hitchhiking" pending datoms), so it
might just be that the fuzzer is finding all permutations of sequences
of a given length.

Still, it is good to know that doing so a least does not violate any of
the internal invariants that already have asserts.