Use slices of different length in fuzzer
Right now, the generated slices do not satisfy the precondition that the
datoms be sorted, so it is expected that counterexamples may be found.
The fuzzer output is not very though, there is only the backtrace, and I
can't even see the panic message.

I should make a generic fuzzing harness, including quickcheck-like
generators that I am now writing in an ad-hoc way. Then I could write
fuzz tests in there, and hook them up to the fuzzer for running, but
once fuzzing finds a counterexample, I want to be able to feed that into
a different binary that then prints the interpretation of the seed in a
human-readable form, followed by violating the assertion, so I can see
the message.