Move most of fuzzing code into a module
This way I can make a program that takes a fuzz artifact as input, and
runs the fuzzer on it, possibly while printing the steps it is taking.
The stack traces I get when libfuzzer exits are not terribly helpful. I
don't get the panic message, and the stack traces look truncated too
sometimes. So it would be helpful to have such a program, where I can
feed the test case into, and then step through it, to inspect it.