Add test case for multilevel inserting tree insert
This is a case caught by the htree_insert fuzzer originally. With the
execution trace, stack trace, and a bit more detailed panic messages,
the problem was easy to pinpoint. When inserting, there might be so many
datoms, that even after splitting, we have more midpoints than fit in a
node.

This commit only adds the test, but does not fix the bug. The test
currently fails.