Fix multilevel tree insertion
This fixes an assertion that was violated when adding more nodes in one
insert, than would fit in two levels of tree. Previously, if a split
produced midpoints to insert at a higher level, we would not consider
that there might be so many midpoints that we need more than one higher
level. This commit fixes that case. All tests now pass again.