Add method to assert that tree invariants hold
This can be used in tests, but especially in fuzzers later: do many many
random operations, and assert that all invariants still hold in between.
If at any point some operation broke an invariant, we will find out
quickly.

The main thing to check about my flavor of the hitchhiker tree, is that
all datoms are sorted. The datoms in each node itself should be sorted,
but also, the datoms in a node should be smaller than the midpoint datom
in the parent that points to it, and greater than the previous midpoint
datom. Also, there are a few things to check about the level of nodes.