Reimplement cursor construction for tree lookup
It is not the most efficient implementation possible: we could do a
binary search through the datoms, rather than a linear scan. And we
could cache the nodes that we looked into, because the iterator will
need them anyway. And we could search for both the upper and lower bound
at once, for the case of iterator construction, to avoid traversing
nodes twice, especially if they are close and both cursors end up
pointing into the same leaf node.

But for now, I will favor something that is simple and which works over
something that is efficient.