Make htree own its comparator
This simplifies things down the road. The instantiator could still pass
in an &'a DatomOrd as Cmp type, so we lose no flexibility. But now the
instantiator is no longer required to have some instance lying around
elsewhere to pass a reference to. This makes the tree more
self-contained. This matters for putting multiple of them in one
database. We can either push the requirement up there too, but then
Database would grow many type parameters, and constructing one would be
inconvenient. The code would be fast due to monomorphization, but
compilation would be slow, and abstractions would leak. I think I prefer
the database to be more self-contained, and having the tree own the
comparator is one step towards that.