Split Store in a Read and Write trait
There is no reason that these should come together. For queries, we only
need to be able to do reads. For meaningful writes we do need reads as
well, but that capability is only required in limited places.

One realization I had, was that it will probably be easiest to not keep
instances of the trees around all the time. That would lead to ownership
conflicts when reading and writing. Rather, I should construct the trees
on demand. They are cheap anyway: all the data is in the store. What a
tree really is, is one page id for the root, a comparator (which is at
this point zero-sized, but might contain a pointer to the value heap
later), and a pointer to the store. So from one store, I can construct
multiple trees that share the underlying store, as long as they are used
for reading only. When I want to write, I need to make sure that only
one tree exists at a time, update it, let it go out of scope, and
construct the next one to apply updates to. This sounds tedious, but it
is also a very neat way of avoiding invalid pointers. For the main
implementation I want to use immutable mmaps anyway, but for testing, I
use the Vec-backed MemoryStore, which may reallocate. So we really
cannot have views into the store through a tree, and write to it at the
same time.

Once again a nice example of how Rust guides you to better design.