Make store::Write imply store::Read
And rename them to store::StoreMut and store::Store, analogous to how
there are Fn and FnMut in the standard library.

In all the cases where we need to be able to write to the store, we also
need to be able to read, because that is how the tree updates work: read
the current node, and write back a modified version. So having the
readable/mutable distinction is a bit more natural, and it simplifies
the type signatures too.