Give database a store, parametrize over store type
The plan is to get rid of the current owned BTreeSet indexes, and
replace them with on-demand constructed indexes that are backed by a
store. So I need a store. And so I need a concrete store type.

I am a bit torn about doing this: adding a type parameter is invasive,
it leaks *everywhere*. And it is not strictly required: one vtable would
not be so bad, especially because the store calls are granular and
infrequent (in comparison to, say, reading datoms from a page). Instead
of adding a type parameter, I could just give the database a boxed trait
object.

However, with the type parameter, it is possible to separate reading and
writing very strictly. All operations that only read, require nothing
more from the store than just being able to read. Only inserts would
require a StoreMut. And that is quite elegant as well. Only the
transactor would need a writable store, and you can tell from the types
that queries will only read from the store. With a trait object,
constructing a read-only database would still require passing a writable
store.

So now the question is: is this a premature optimization that makes all
the code ugly, and a leaky abstraction, or is it a good design decision?
It is not yet obvious to me which one it is. I will go with the type
parameter for now. It can be removed later if it turns out to be very
tedious.