Add other indexes, use them in support functions
They are not yet used for query.

Do not add the VAET index. I added it because Datomic has it, but so
far I have nod needed it, and I don't think I will, because all queries
have attributes fixed in the query, so any query that VAET would be
useful for, could also be answered by AVET: if we want to query "which
entities have this value for a given attribute", then we know both the
value and the attribute, so either index would do.

If VAET does turn out to be useful to me later, it will be easy to add.