Use htree indexes in database for query
Queries are now powered by my hitchhiker tree, not by the BTreeSet any
more. This means that we do lose the writes that happen in the first
transaction, as writes have not yet been ported. This will be addressed
soon.

This was quite an adventure to get the lifetimes right, and I am not
entirely satisfied with the end result. It is a bit messy. I suspect
that a better design exists, and I have not yet discovered it, but I
have a hunch of what it might look like.

One thing that stands out at this point, is that tree insertion (and
the splits and flush that come with it) have nothing to do at all with
iteration and range queries. Perhaps I should split those two up.

Another thing that happened now is that both htree::HTree and
htree::Iter are lightweight structs that wrap a store. Iter creation
now goes via the HTree first, but there is little need for that now. It
might be better to separate them completely. That would also solve the
problem where insert may update the root page id, and the database would
need to get it back later.