Make insertion insert into the htrees
This is a bit rough, many results are ignored. Inserting a single datom
at a time is bad for the htree, it is much better suited for batch
inserts, and I should change the interface to deal with this. First
build up a vector of datoms to insert, then do one insert.

Now the new entities get stored, but the main program still does not
print them, because it uses the old tree roots. Which is kind of neat in
a way, because it demonstrates that immutability pays off, and that we
can just query an old version of the index even if there is new data.
Note that this is unrelated to a scan that ignores newer datoms, which
should be implemented anyway. Scanning an older index would be better,
if possible, because it avoids the new datoms entirely.