Remove orderable datom wrappers
These were used because the B-tree set in the standard library uses the
type's Ord instance, and there can only be one Ord. For the different
indexes, we need to order Datom in different ways. So we needed a
wrapper type. But the new HTree takes an external comparator, so there
is no need to wrap the datom.