Bring file docs a bit more up to date
The value heap I still plan to add, it just has not been implemented
yet. The journal I am less sure about. Perhaps I'll just add a
transaction ordered index and use that as a substitute. Flushing to new
tree nodes on every transaction may be wasteful, but my target is to get
something working first, and flushing is not a big problem yet.