Prefix variables with $ in example
This makes the examples a lot noisier, but it should also help to
distinguish literal attribute names from variables.

Perhaps this is the wrong thing to do, and I should instead use a sigil
for attributes that the database resolves. For example, instead of

    where
      $i issue.title "Lifespan is too short"
      $i issue.status $status
    select
      $status

we might write

    where
      i @issue.title "Lifespan is too short"
      i @issue.title status
    select
      status

It is a bit cleaner, but it might be alient too. At least everybody has
seen PHP with its $variables, and Rust's macros also use that sigil, so
it should come as no surprise.