Fix type errors in Haskell client library
I hesitated a bit between a Noblit type that implements Monad, and a
MonadNoblit class. But for the latter, concrete instances would need a
way to handle errors, so it would turn into something like

    instance (MonadIO m, MonadError Error m) => MonadNoblit m

Which maybe isn't so bad after all. With a Noblit type, I would go with
the free monad aproach, and a `run` interpreter for it that returns the
error. It is a bit more explicit, but it does not allow freely mixing
multiple monads (if you are not fully buying in to free monads, and in
the Haskell ecosystem transformers have won, not free monads), whereas
MonadNoblit would ... hmm, I'll have to think about it.