Improve type safety of constant ids
By not sharing the same offset type for integer constants and byte
constants, we can make it harder to confuse the two. That means that if
you got an integer constant from somewhere, it will be hard to
accidentally get a string constant from it and vice versa.

I don't really like the names CidInt and CidBytes, but I don't have any
better ideas either at this point.