Introduce QueryEngine to track temporary values
QueryEngine is not really the best name for it, but for lack of
something better ...

The engine provides a view of the database, together with a stack pool
where temporary values can be stored. The indexes will then be able to
compare temporary values against values stored in the database.

We could go a step further and parametrize values and the query plan
over the lifetime of the query engine, but I think that would become
tedious; tracking it at runtime, possibly with a generation number to
detect bugs early, seems like the way to go at this point.