Put StackPool under source control
I forgot to commit this file, how embarrassing ...

I can put it back and rebase all commits, but I don't have the history
for the file, so there is no guarantee that that would work. If you want
to compile an old commit, add this file with it. I am sorry.