Add function to generate interesting fuzz input
I can generate an input that creates a tree of depth 3 easily, but the
fuzzer does not discover it, even when I generate an input that is only
one mutation away from a depth-3 input. This is a bit worrisome, because
the fuzzer may not be finding some interesting inputs. Needs further
investigation.