Make htree_insert_value fuzzer command-based
So the fuzzer were all running terribly. I suspect this was because
length-prefixed inputs are bad for the fuzzer to recombine and mutate.
So I switched it up, to use non-contiguous chunks of the input data,
with 4-bit length prefixes, to make the effects of mutation more local.
But that did not help so much either.

Then I realized: I should be treating the input as a command-stream, not
length-prefixed. So now I switched up the htree_insert_value fuzzer, but
I should take on the other one too. We read one byte from input, and
that value decides what to do; it's a command. An insert command then
reads some extra data (a length, and then the slice). So now the fuzz
input is something like an instruction stream, with very domain-specific
instructions.

The advantage of this instruction stream, is that you can insert bytes
in the middle, or cross it over with another instruction stream, and the
result will be a similar instruction stream with high probability, so
the mutations that the fuzzer makes are effective. With length-prefix
this was not the case: if you change the length value, that changes the
meaning of everything after that data slice. Inserts don't work
naturally there, but in the instruction stream, they do.

The fuzzer at least looks much more active this way. It's still running,
but I am somewhat confident that it will be able to find instructions to
construct a tree of height 4.