Add a toy query parser
I want to start writing a test suite where I can write queries (both
select and assert), and the expected output, and build a somewhat
readable suite of golden tests.

It turns out that, even in the Haskell client I'm working on, having to
pre-declare all the variables you want to use in the query gets tedious
quickly. I suspect that Datomic in Clojure uses symbols for this reason:
it gets you stable identifiers without having to pre-declare all
variables, and without awkward string literals. I could do string
literals in Haskell, but then you lose the benefits of the typechecker.

In the mean time, I think that my made-up query language is a bit nicer
to write than what you can do in any general-purpose programming
language (although I have to admit that I don't know any Lisp well
enough to judge whether it would be appropriate). So I'll write a toy
parser for my ad-hoc query language, such that I can use it in tests.

The parser converts identifiers into integers (which Noblit expects).
This is easily done in a program, but tedious (and unreadable) for
humans. The parser also handles string escaping through Python's parser.
Then we get out a data structure that is quite similar to what Noblit
expects on the Rust side. The next step is to serialize the structure
and to read it in Rust. The point of the Python parser is that it's
easier to write such a toy parser in Python than it is to do so in Rust.
The Python parser is not meant to handle edge cases or do proper error
handling. If I ever add a proper parser, it would probably be written in
Rust. But for now, this will get me a way to write queries quickly.