Add test runner for golden tests
Now I can run

    prove --exec golden/run.py golden

And it will take the .t files in the "golden" directory, run the queries
in them, compare the output against the expectation, and print in TAP
compliant format whether the output was as expected. So now I can start
writing golden tests. Also, I can almost remove the "main" binary, as
the combination of golden/parse.py and the "execute" binary is more
flexible. Only I don't have a way yet to insert new data from tests.
I should extend all of this with support for asserting datoms.