Add Nix-based build environment, use it on CI
I am tired of having to fix missing packages, or having the wrong
versions of packages, on CI all the time. It is fixable, but it turns
into a very specific configuration that works for Travis' interpretation
of Ubuntu, and it is getting out of hand bit by bit.

A separate problem is reproducibility: if I check out a commit two years
from now, I still want to be able to build and run Noblit. This implies
pinning everything -- dependencies, but also the entire toolchain. Parts
are already pinned: Rustup pins Cargo and Rustc, Stack pins GHC and
Haskell dependencies. Now I add Nix to pin Rustup and Stack, and also
some other development dependencies such as Python that are used to run
the tests.

So Nix solves the reproducibility problem. It also isolates
dependencies, which means I can use the same build environment here, and
on CI. That should make CI a lot easier to debug. And it also solves the
toolchain problem neatly: if the Nix environment works locally, then it
will work on CI too.

An additional advantage is that this reduces vendor lock-in on the CI
provider. Anything that can run Nix can now create an environment with
the right tools to build Noblit.